WIP: Puyolib: Basic custom skin support
Allows the peer to accept and send custom folders to get animations
and background images from in an online match. There is currently
no way to actually set a custom skin string. This will likely be
a custom UI element in the ChatWindow widget instead of being a
part of the game itself.

WARNING: This is terribly unsafe and only serves as POC. DO NOT USE