Puyolib, Client - Add support for archive packages
Allows Puyolib to use assets provided by a ZIP archive containing the
file structure of the Assets/ folder without a parent folder (that is, the
root of the archive contains an User folder etc.) Wrongly-formatted archives
should get filtered and unused but might result in a crash.

Compatibility with Windows is untested, retrieving Client assets (such as the
logo) is not possible, will be easier to bake into the executable.